WIP. Begin efforts to rework into a single table. Refactor some parts of the code. Add some TODOs xibosignage/xibo#1320
