+++ install/steps/135.json
@@ -1,17 +1,17 @@
 {
   "dbVersion": 135,
-  "appVersion": "1.8.3",
+  "appVersion": "1.8.4",
   "steps": [
     {
-      "step": "Create a new Table for remote Dataset settings",
-      "action": "CREATE TABLE IF NOT EXISTS `datasetremote` (`rdId` bigint(20) NOT NULL AUTO_INCREMENT,`DataSetID` int(11) NOT NULL, `method` ENUM('GET','POST') NOT NULL, `uri` varchar(250) NOT NULL, `postData` TEXT DEFAULT NULL, `authentication` ENUM('none','plain','basic','digest') NOT NULL, `username` varchar(100) DEFAULT NULL, `password` varchar(250) DEFAULT NULL, `refreshRate` int(11) DEFAULT 86400, `clearRate` int(11) DEFAULT 0, `runsAfter` int(11) DEFAULT NULL, `dataRoot` varchar(250) DEFAULT NULL, `lastSync` int(11) NOT NULL DEFAULT 0, `summarize` varchar(10) DEFAULT NULL, `summarizeField` varchar(250) DEFAULT NULL, PRIMARY KEY (`rdId`)) ENGINE=InnoDB  DEFAULT CHARSET=utf8 AUTO_INCREMENT=1;"
+      "step": "Add new columns for Remote DataSets",
+      "action": "ALTER TABLE `dataset` ADD COLUMN `isRemote` TINYINT(4) NOT NULL DEFAULT 0;ALTER TABLE `dataset` ADD COLUMN `method` ENUM('GET','POST') DEFAULT NULL;ALTER TABLE `dataset` ADD COLUMN `uri` varchar(250) DEFAULT NULL;ALTER TABLE `dataset` ADD COLUMN `postData` TEXT DEFAULT NULL;ALTER TABLE `dataset` ADD COLUMN `authentication` ENUM('none','plain','basic','digest') DEFAULT NULL;ALTER TABLE `dataset` ADD COLUMN `username` varchar(100) DEFAULT NULL;ALTER TABLE `dataset` ADD COLUMN `password` varchar(250) DEFAULT NULL;ALTER TABLE `dataset` ADD COLUMN `refreshRate` int(11) DEFAULT 86400;ALTER TABLE `dataset` ADD COLUMN `clearRate` int(11) DEFAULT 0;ALTER TABLE `dataset` ADD COLUMN `runsAfter` int(11) DEFAULT NULL;ALTER TABLE `dataset` ADD COLUMN `dataRoot` varchar(250) DEFAULT NULL;ALTER TABLE `dataset` ADD COLUMN `lastSync` int(11) NOT NULL DEFAULT 0;ALTER TABLE `dataset` ADD COLUMN `summarize` varchar(10) DEFAULT NULL;ALTER TABLE `dataset` ADD COLUMN `summarizeField` varchar(250) DEFAULT NULL;"
     },
     {
-      "step": "Create a new ColumnType",
+      "step": "Add a column type for remote DataSets",
       "action": "INSERT INTO `datasetcolumntype` SET dataSetColumnTypeId=3, dataSetColumnType='Remote';"
     },
     {
-      "step": "Create a new Column in DataSetColumn",
+      "step": "Add a new column to the DataSet column table for remote DataSets",
       "action": "ALTER TABLE `datasetcolumn` ADD `RemoteField` VARCHAR(250) DEFAULT NULL AFTER `formula`;"
     },
     {
