+++ lib/Controller/DataSet.php
@@ -1,7 +1,8 @@
 <?php
 /*
  * Xibo - Digital Signage - http://www.xibo.org.uk
- * Copyright (C) 2011-2013 Daniel Garner
+ * Copyright (C) 2011-2017 Spring Signage Ltd
+ * contributions by LukyLuke aka Lukas Zurschmiede - https://github.com/LukyLuke
  *
  * This file is part of Xibo.
  *
@@ -21,7 +22,9 @@
 namespace Xibo\Controller;
 
 use Xibo\Exception\AccessDeniedException;
+use Xibo\Exception\InvalidArgumentException;
 use Xibo\Exception\NotFoundException;
+use Xibo\Exception\XiboException;
 use Xibo\Factory\DataSetColumnFactory;
 use Xibo\Factory\DataSetFactory;
 use Xibo\Helper\DataSetUploadHandler;
@@ -164,7 +167,6 @@ class DataSet extends Base
             $dataSet->load();
 
             if ($user->checkEditable($dataSet)) {
-                $isRemote = $this->dataSetFactory->isRemoteDataSet($dataSet->dataSetId);
 
                 // View Data
                 $dataSet->buttons[] = array(
@@ -186,7 +188,7 @@ class DataSet extends Base
                 $dataSet->buttons[] = ['divider' => true];
 
                 // Import DataSet
-                if (!$isRemote) {
+                if ($dataSet->isRemote !== 1) {
                     $dataSet->buttons[] = array(
                         'id' => 'dataset_button_import',
                         'class' => 'dataSetImportForm',
@@ -198,7 +200,7 @@ class DataSet extends Base
                 // Copy
                 $dataSet->buttons[] = array(
                     'id' => 'dataset_button_copy',
-                    'url' => $this->urlFor('dataSet.copy' . ($isRemote ? '.remote' : '') . '.form', ['id' => $dataSet->dataSetId]),
+                    'url' => $this->urlFor('dataSet.copy.form', ['id' => $dataSet->dataSetId]),
                     'text' => __('Copy')
                 );
 
@@ -208,7 +210,7 @@ class DataSet extends Base
                 // Edit DataSet
                 $dataSet->buttons[] = array(
                     'id' => 'dataset_button_edit',
-                    'url' => $this->urlFor('dataSet.edit' . ($isRemote ? '.remote' : '') . '.form', ['id' => $dataSet->dataSetId]),
+                    'url' => $this->urlFor('dataSet.edit.form', ['id' => $dataSet->dataSetId]),
                     'text' => __('Edit')
                 );
             }
@@ -217,7 +219,7 @@ class DataSet extends Base
                 // Delete DataSet
                 $dataSet->buttons[] = array(
                     'id' => 'dataset_button_delete',
-                    'url' => $this->urlFor('dataSet.delete' . ($isRemote ? '.remote' : '') . '.form', ['id' => $dataSet->dataSetId]),
+                    'url' => $this->urlFor('dataSet.delete.form', ['id' => $dataSet->dataSetId]),
                     'text' => __('Delete')
                 );
             }
@@ -247,6 +249,7 @@ class DataSet extends Base
     {
         $this->getState()->template = 'dataset-form-add';
         $this->getState()->setData([
+            'dataSets' => $this->dataSetFactory->query(),
             'help' => $this->getHelp()->link('DataSet', 'Add')
         ]);
     }
@@ -292,6 +295,8 @@ class DataSet extends Base
      *      )
      *  )
      * )
+     *
+     * @throws XiboException
      */
     public function add()
     {
@@ -299,7 +304,24 @@ class DataSet extends Base
         $dataSet->dataSet = $this->getSanitizer()->getString('dataSet');
         $dataSet->description = $this->getSanitizer()->getString('description');
         $dataSet->code = $this->getSanitizer()->getString('code');
+        $dataSet->isRemote = $this->getSanitizer()->getCheckbox('isRemote');
         $dataSet->userId = $this->getUser()->userId;
+
+        // Fields for remote
+        if ($dataSet->isRemote === 1) {
+            $dataSet->method = $this->getSanitizer()->getString('method');
+            $dataSet->uri = $this->getSanitizer()->getString('uri');
+            $dataSet->postData = $this->getSanitizer()->getString('postData');
+            $dataSet->authentication = $this->getSanitizer()->getString('authentication');
+            $dataSet->username = $this->getSanitizer()->getString('username');
+            $dataSet->password = $this->getSanitizer()->getString('password');
+            $dataSet->refreshRate = $this->getSanitizer()->getInt('refreshRate');
+            $dataSet->clearRate = $this->getSanitizer()->getInt('clearRate');
+            $dataSet->runsAfter = $this->getSanitizer()->getInt('runsAfter');
+            $dataSet->dataRoot = $this->getSanitizer()->getString('dataRoot');
+            $dataSet->summarize = $this->getSanitizer()->getString('summarize');
+            $dataSet->summarizeField = $this->getSanitizer()->getString('summarizeField');
+        }
 
         // Also add one column
         $dataSetColumn = $this->dataSetColumnFactory->createEmpty();
@@ -339,6 +361,7 @@ class DataSet extends Base
         $this->getState()->template = 'dataset-form-edit';
         $this->getState()->setData([
             'dataSet' => $dataSet,
+            'dataSets' => $this->dataSetFactory->query(),
             'help' => $this->getHelp()->link('DataSet', 'Edit')
         ]);
     }
@@ -387,6 +410,8 @@ class DataSet extends Base
      *      @SWG\Schema(ref="#/definitions/DataSet")
      *  )
      * )
+     *
+     * @throws XiboException
      */
     public function edit($dataSetId)
     {
@@ -398,6 +423,23 @@ class DataSet extends Base
         $dataSet->dataSet = $this->getSanitizer()->getString('dataSet');
         $dataSet->description = $this->getSanitizer()->getString('description');
         $dataSet->code = $this->getSanitizer()->getString('code');
+        $dataSet->isRemote = $this->getSanitizer()->getCheckbox('isRemote');
+
+        if ($dataSet->isRemote === 1) {
+            $dataSet->method = $this->getSanitizer()->getString('method');
+            $dataSet->uri = $this->getSanitizer()->getString('uri');
+            $dataSet->postData = $this->getSanitizer()->getString('postData');
+            $dataSet->authentication = $this->getSanitizer()->getString('authentication');
+            $dataSet->username = $this->getSanitizer()->getString('username');
+            $dataSet->password = $this->getSanitizer()->getString('password');
+            $dataSet->refreshRate = $this->getSanitizer()->getInt('refreshRate');
+            $dataSet->clearRate = $this->getSanitizer()->getInt('clearRate');
+            $dataSet->runsAfter = $this->getSanitizer()->getInt('runsAfter');
+            $dataSet->dataRoot = $this->getSanitizer()->getString('dataRoot');
+            $dataSet->summarize = $this->getSanitizer()->getString('summarize');
+            $dataSet->summarizeField = $this->getSanitizer()->getString('summarizeField');
+        }
+
         $dataSet->save();
 
         // Return
@@ -410,7 +452,8 @@ class DataSet extends Base
 
     /**
      * DataSet Delete
-     * * @param int $dataSetId
+     * @param int $dataSetId
+     * @throws XiboException
      */
     public function deleteForm($dataSetId)
     {
@@ -452,6 +495,8 @@ class DataSet extends Base
      *      description="successful operation"
      *  )
      * )
+     *
+     * @throws XiboException
      */
     public function delete($dataSetId)
     {
@@ -460,11 +505,9 @@ class DataSet extends Base
         if (!$this->getUser()->checkDeleteable($dataSet))
             throw new AccessDeniedException();
 
-        $this->getLog()->debug('Delete data flag = ' . $this->getSanitizer()->getCheckbox('deleteData') . '. Params = ' . var_export($this->getApp()->request()->params(), true));
-
         // Is there existing data?
         if ($this->getSanitizer()->getCheckbox('deleteData') == 0 && $dataSet->hasData())
-            throw new \InvalidArgumentException(__('There is data assigned to this data set, cannot delete.'));
+            throw new InvalidArgumentException(__('There is data assigned to this data set, cannot delete.'), 'dataSetId');
 
         // Otherwise delete
         $dataSet->delete();
@@ -540,6 +583,8 @@ class DataSet extends Base
      *      @SWG\Schema(ref="#/definitions/DataSet")
      *  )
      * )
+     *
+     * @throws XiboException
      */
     public function copy($dataSetId)
     {
@@ -617,6 +662,9 @@ class DataSet extends Base
      *      description="successful operation"
      *  )
      * )
+     *
+     * @throws XiboException
+     * @throws \Exception
      */
     public function import($dataSetId)
     {
@@ -789,5 +837,232 @@ class DataSet extends Base
             'httpStatus' => 204,
             'message' => sprintf(__('Imported JSON into %s'), $dataSet->dataSet)
         ]);
+    }
+
+    /**
+     * Sends out a TestRequst and returns the Data as JSON to the Client so it can be shown in the Dialog
+     */
+    public function testRemoteRequest() {
+        $dataSet = $this->dataSetFactory->createEmptyRemote();
+        $dataSet->dataSet = $this->getSanitizer()->getString('dataSet');
+        $dataSet->method = $this->getSanitizer()->getString('method');
+        $dataSet->uri = $this->getSanitizer()->getString('uri');
+        $dataSet->postData = $this->getSanitizer()->getString('postData');
+        $dataSet->authentication = $this->getSanitizer()->getString('authentication');
+        $dataSet->username = $this->getSanitizer()->getString('username');
+        $dataSet->password = $this->getSanitizer()->getString('password');
+        $dataSet->dataRoot = $this->getSanitizer()->getString('dataRoot');
+
+        $data = $this->dataSetFactory->callRemoteService($dataSet);
+
+        // Return
+        $this->getState()->hydrate([
+            'message' => sprintf(__('Run Test-Request for %s on %s'), $dataSet->dataSet, $dataSet->getCurlParams()[CURLOPT_URL]),
+            'id' => $dataSet->dataSetId,
+            'data' => $data->entries[0]
+        ]);
+    }
+
+    /**
+     * Tries to process received Data against the configured DataSet with all Columns
+     *
+     * @param \Xibo\Entity\DataSetRemote $dataSet The RemoteDataset to process
+     * @param \stdClass $results A simple Object with one Property 'entries' which contains all results
+     */
+    public function processResults(\Xibo\Entity\DataSetRemote $dataSet, \stdClass $results) {
+        if (property_exists($results, 'entries') && is_array($results->entries)) {
+            foreach ($results->entries as $result) {
+                $this->process($dataSet, $result);
+            }
+        }
+    }
+
+    /**
+     * Tries to process received Data against the configured DataSet with all Columns
+     *
+     * @param \Xibo\Entity\DataSetRemote $dataSet The RemoteDataset to process
+     * @param array The JSON received from the remote endpoint
+     */
+    private function process(\Xibo\Entity\DataSetRemote $dataSet, array $result) {
+        // Remote Data has to have the configured DataRoot which has to be an Array
+        $data = $this->getDataRootFromResult($dataSet->dataRoot, $result);
+        if (($data != null) && is_array($data)) {
+            $columns = $this->dataSetColumnFactory->query(null, ['dataSetId' => $dataSet->dataSetId]);
+            $entries = [];
+
+            // First process each entry form the remote and try to map the values to the configured columns
+            foreach($data as $k => $entry) {
+                if (is_array($entry) || is_object($entry)) {
+                    $entries[] = $this->processEntry($dataSet, (array) $entry, $columns);
+                } else {
+                    $message = sprintf(__('DataSet \'%s\' failed: DataRoot \'%s\' contains data which are not arrays and not objects.'), $dataSet->dataSet, $dataSet->dataRoot);
+                    break;
+                }
+            }
+
+            // If there is a Consilidation-Function, use the Data against it
+            $entries = $this->consolidateEntries($dataSet, $entries, $columns);
+
+            // Finally add each entry as a new Row in the DataSet
+            foreach ($entries as $entry) {
+                $dataSet->addRow($entry);
+            }
+        } else {
+            $message = sprintf(__('DataSet \'%s\' missconfigured: DataRoot \'%s\' is not an Array.'), $dataSet->dataSet, $dataSet->dataRoot);
+        }
+
+        // Return
+        $this->getState()->hydrate([
+            'message' => $message,
+            'id' => $dataSet->dataSetId
+        ]);
+    }
+
+    /**
+     * Process the RemoteResult to get the main DataRoot value which can be stay in a structure as well as the values
+     *
+     * @param String Chuns splitted by a Dot where the main entries are hold
+     * @param array The Value from the remote request
+     * @return array The Data hold in the configured dataRoot
+     */
+    private function getDataRootFromResult($dataRoot, array $result) {
+        if (empty($dataRoot)) {
+            return $result;
+        }
+        $chunks = explode('.', $dataRoot);
+        $entries = $this->getFieldValueFromEntry($chunks, $result);
+        return $entries[1];
+    }
+
+    /**
+     * Process a single Data-Entry form the remote system and map it to the configured Columns
+     *
+     * @param \Xibo\Entity\DataSetRemote $dataSet The DataSet which is processed currently
+     * @param array $entry The Data from the remte system
+     * @param array $columns The configured Columns form the current DataSet
+     * @return array The processed $entry as a List of Fields from $columns
+     */
+    private function processEntry(\Xibo\Entity\DataSetRemote $dataSet, array $entry, array $columns) {
+        $result = [];
+
+        foreach ($columns as $k => $column) {
+            if (($column->remoteField != null) && ($column->remoteField != '')) {
+                $dataTypeId = $column->dataTypeId;
+
+                // The Field may be a Date, timestamp or a real field
+                if ($column->remoteField == '{{DATE}}') {
+                    $value = [0, date('Y-m-d')];
+
+                } else if ($column->remoteField == '{{TIMESTAMP}}') {
+                    $value = [0, time()];
+
+                } else {
+                    $chunks = explode('.', $column->remoteField);
+                    $value = $this->getFieldValueFromEntry($chunks, $entry);
+                }
+
+                // Only add it to the result if we where able to process the field
+                if (($value != null) && ($value[1] != null)) {
+                    switch ($dataTypeId) {
+                        case 2:
+                            $result[$column->heading] = $this->getSanitizer()->double($value[1]);
+                            break;
+                        case 3:
+                            $result[$column->heading] = $this->getDate()->getLocalDate(strtotime($value[1]));
+                            break;
+                        case 5:
+                            $result[$column->heading] = $this->getSanitizer()->int($value[1]);
+                            break;
+                        default:
+                            $result[$column->heading] = $this->getSanitizer()->string($value[1]);
+                    }
+                }
+            }
+        }
+
+        return $result;
+    }
+
+    /**
+     * Returns the Value of the remote DataEntry based on the remoteField definition splitted into chunks
+     *
+     * This function is recursive, so be sure you remove the first value from chunks and pass it in again
+     *
+     * @param array List of Chunks which interprets the FieldNames in the actual DataEntry
+     * @param array $entry Current DataEntry
+     * @return array of the last FieldName and the corresponding value
+     */
+    private function getFieldValueFromEntry(array $chunks, array $entry) {
+        $value = null;
+        $key = array_shift($chunks);
+
+        if (($entry instanceof \StdClass) && property_exists($entry, $key)) {
+            $value = $entry->{$key};
+        } else if (array_key_exists($key, $entry)) {
+            $value = $entry[$key];
+        }
+
+        if (($value != null) && (count($chunks) > 0)) {
+            return $this->getFieldValueFromEntry($chunks, (array) $value);
+        }
+
+        return [ $key, $value ];
+    }
+
+    /**
+     * Consolidates all Entries by the defined Function in the DataSet
+     *
+     * This Method *sums* or *counts* all same entries and returns them.
+     * If no consolidation function is configured, nothing is done here.
+     *
+     * @param \Xibo\Entity\DataSetRemote $dataSet the current DataSet
+     * @param array $entries All processed entries which may be consolidated
+     * @param array $columns The columns form this DataSet
+     * @return \Slim\Helper\Set which contains all Entries to be added to the DataSet
+     */
+    private function consolidateEntries(\Xibo\Entity\DataSetRemote $dataSet, array $entries, array $columns) {
+        if ((count($entries) > 0) && $dataSet->doConsolidate()) {
+            $consolidated = new \Slim\Helper\Set();
+            $field = $dataSet->getConsolidationField();
+
+            // Get the Field-Heading based on the consolidation field
+            foreach ($columns as $k => $column) {
+                if ($column->remoteField == $dataSet->summarizeField) {
+                    $field = $column->heading;
+                    break;
+                }
+            }
+
+            // Check each entry and consolidate the value form the defined field
+            foreach ($entries as $entry) {
+                if (array_key_exists($field, $entry)) {
+                    $key = $field . '-' . $entry[$field];
+                    $existing = $consolidated->get($key);
+
+                    // Create a new one if there is no currently consolidated field for this value
+                    if ($existing == null) {
+                        $existing = $entry;
+                        $existing[$field] = 0;
+                    }
+
+                    // Consolidate: Summarize, Count, Unknown
+                    if ($dataSet->summarize == 'sum') {
+                        $existing[$field] = $existing[$field] + $entry[$field];
+
+                    } else if ($dataSet->summarize == 'count') {
+                        $existing[$field] = $existing[$field] + 1;
+
+                    } else {
+                        // Unknown consolidation type :?
+                        $existing[$field] = 0;
+                    }
+
+                    $consolidated->set($key, $existing);
+                }
+            }
+
+            return $consolidated;
+        }
+        return new \Slim\Helper\Set($entries);
     }
 }
