+++ views/dataset-page.twig
@@ -13,7 +13,6 @@
 {% block actionMenu %}
     <ul class="nav nav-pills pull-right">
         <li class="btn btn-success btn-xs"><a class="XiboFormButton btns" title="{% trans "Add a new DataSet" %}" href="{{ urlFor("dataSet.add.form") }}"> <i class="fa fa-plus-circle" aria-hidden="true"></i> {% trans "Add DataSet" %}</a></li>
-        <li class="btn btn-success btn-xs"><a class="XiboFormButton btns" title="{% trans "Add a new Remote DataSet" %}" href="{{ urlFor("dataSet.add.remote.form") }}"> <i class="fa fa-plus-circle" aria-hidden="true"></i> {% trans "Add Remote DataSet" %}</a></li>
     </ul>
 {% endblock %}
 
@@ -38,6 +37,7 @@
                                 <th>{% trans "Name" %}</th>
                                 <th>{% trans "Description" %}</th>
                                 <th>{% trans "Code" %}</th>
+                                <th>{% trans "Remote?" %}</th>
                                 <th>{% trans "Owner" %}</th>
                                 <th>{% trans "Permissions" %}</th>
                                 <th></th>
@@ -71,6 +71,10 @@
                 { "data": "dataSet", "render": dataTableSpacingPreformatted },
                 { "data": "description" },
                 { "data": "code" },
+                {
+                    "data": "isRemote",
+                    "render": dataTableTickCrossColumn
+                },
                 { "data": "owner" },
                 { "data": "groupsWithPermissions" },
                 {
@@ -180,6 +184,12 @@
             });
         });
         table.on('processing.dt', dataTableProcessing);
+
+        function XiboDatasetRemoteTestRequest($form) {
+            XiboRemoteRequest("{{ urlFor("dataSet.test.remote") }}", $("#dataSetRemoteForm").serializeObject(), function(response) {
+                $("#datasetRemoteTestRequestResult").html('<pre>' + JSON.stringify(response.data, null, 3) + '</pre>');
+            });
+        }
     </script>
 {% endblock %}
